+++ src/main/java/org/owasp/webgoat/container/users/UserTracker.java
@@ -16,7 +16,7 @@ import org.owasp.webgoat.container.lessons.Lesson;
  * <p>
  *
  * <p>This file is part of WebGoat, an Open Web Application Security Project utility. For details,
- * please see http://www.owasp.org/
+ * please see https://www.owasp.org/
  *
  * <p>Copyright (c) 2002 - 2014 Bruce Mayhew
  *
@@ -37,7 +37,7 @@ import org.owasp.webgoat.container.lessons.Lesson;
  * <p>Source for this application is maintained at https://github.com/WebGoat/WebGoat, a repository
  * for free software projects.
  *
- * @author Bruce Mayhew <a href="http://code.google.com/p/webgoat">WebGoat</a>
+ * @author Bruce Mayhew <a href="https://code.google.com/p/webgoat">WebGoat</a>
  * @version $Id: $Id
  * @since October 29, 2003
  */